ClouderaSetup/OnPremise/Automation/pvcauto/webapp/requirements.txt
Flask==2.2.5
PyYAML>=6.0,<7.0
Cython>=0.29.34
ansible